refactor(console): move migration proxy to /api/migration/[endpoint]
workspaceId is now an optional query param: present -> session required
+ workspace access verified, forwarded with the caller's auth; absent ->
anonymous/teaser mode, forwarded with the service token and gated by the
report accessCode on the ee-api side. Per-IP rate limits unchanged and
now cover anonymous callers too. Wizard client updated to the new paths.
